Fix ClassCastException when enum fields are requested in GraphQL query
Job priority and Task status are simple enums that depict respectively a single
state in Job and Task field. In other words, a given Job cannot have multiple
priorities defined. Same for tasks.

Also, GraphQL Java engine expects to receive a String as enum value for the
result of the data fetcher. Returning an enum is not working and leading to a
null value returned to the user. The reason lies in the fact that GraphQL engine
is checking the value returned by the data fetcher (with equals method) against
the GraphQLEnumType that is defined in the Schema.