Add support for authentication
All requests to the scheduling API now requires a session ID to be defined and
valid. If it is not the case, an exception is raised.

The GraphiQL endpoint has been updated to fetch the session id from local
storage. This way, it will be possible to authenticate using an existing portal.