Gradle projects refactoring
The purpose is to put in dedicated modules different concepts to prevent the
usage of libraries and code that are not related.

GraphQL argument, field and type names have also been extracted in dedicated
classes to be able to share them between the server and the client
implementations. This prevents code duplication and ease, again, maintenance.

Source code has also been reformatted.