Refactor GraphQL client and beans modules
The purpose is to put all beans in dedicated and reusable modules.

New modules are introduced for GWT compatibility.

Beans, especially input beans, must not rely on Lombok since GWT
requires the source code of the Java classes in order to compile it to
Javascript.

References to String#format have been removed since GWT emulation is not
supported for this method.